+++ test/multiverse/suites/sidekiq/sidekiq_ignore_retry_errors_disabled_test.rb
@@ -4,6 +4,9 @@
 
 require_relative 'sidekiq_test_helpers'
 
+# On startup, Sidekiq instrumentation registers error and death handlers
+# based on the value of the 'sidekiq.ignore_retry_errors'. Because of this,
+# we need to have separate enabled/disabled test suites to test both cases.
 class SidekiqIgnoreRetryErrorsTest < Minitest::Test
   include SidekiqTestHelpers
 
@@ -50,69 +53,6 @@ class SidekiqIgnoreRetryErrorsTest < Minitest::Test
       'Expected NewRelic error_handler to be registered when sidekiq.ignore_retry_errors is false'
   end
 
-  def test_error_handlers_not_registered_when_sidekiq_ignore_retry_errors_is_true
-    with_config(:'sidekiq.ignore_retry_errors' => true) do
-      # TODO: MAJOR VERSION - remove this when Sidekiq v5 is no longer supported
-      skip 'Test requires Sidekiq v6+' unless Sidekiq::VERSION.split('.').first.to_i >= 6
-
-      config = if Sidekiq::VERSION.split('.').first.to_i >= 7
-        Sidekiq.default_configuration
-      else
-        Sidekiq
-      end
-
-      error_handlers = if config.respond_to?(:error_handlers)
-        config.error_handlers
-      else
-        config[:error_handlers] || []
-      end
-
-      nr_error_handler_found = error_handlers.any? do |handler|
-        handler.is_a?(Proc) && handler.source_location&.first&.include?('newrelic')
-      end
-
-      refute nr_error_handler_found,
-        'Expected NewRelic error_handler to NOT be registered when sidekiq.ignore_retry_errors is true'
-    end
-  end
-
-  def test_death_handlers_registered_when_sidekiq_ignore_retry_errors_is_true
-    with_config(:'sidekiq.ignore_retry_errors' => true) do
-      # TODO: MAJOR VERSION - remove this when Sidekiq v5 is no longer supported
-      skip 'Test requires Sidekiq v6+' unless Sidekiq::VERSION.split('.').first.to_i >= 6
-
-      config = if Sidekiq::VERSION.split('.').first.to_i >= 7
-        Sidekiq.default_configuration
-      else
-        Sidekiq
-      end
-
-      puts '**************************'
-      puts "Sidekiq version: #{Sidekiq::VERSION}"
-      puts "config: #{config.inspect}"
-
-      death_handlers = if config.respond_to?(:death_handlers)
-        config.death_handlers
-      else
-        config[:death_handlers] || []
-      end
-
-      puts "death_handlers: #{death_handlers.inspect}"
-
-      nr_death_handler_found = death_handlers.any? do |handler|
-        handler.is_a?(Proc) && handler.source_location&.first&.include?('newrelic')
-      end
-
-      puts "nr_death_handler_found: #{nr_death_handler_found}"
-
-      puts "sidekiq.ignore_retry_errors: #{NewRelic::Agent.config[:'sidekiq.ignore_retry_errors']}"
-      puts '**************************'
-
-      assert nr_death_handler_found,
-        'Expected NewRelic death_handler to be registered when sidekiq.ignore_retry_errors is true'
-    end
-  end
-
   def test_death_handlers_not_registered_when_sidekiq_ignore_retry_errors_is_false
     # TODO: MAJOR VERSION - remove this when Sidekiq v5 is no longer supported
     skip 'Test requires Sidekiq v6+' unless Sidekiq::VERSION.split('.').first.to_i >= 6
@@ -135,15 +75,5 @@ class SidekiqIgnoreRetryErrorsTest < Minitest::Test
 
     refute nr_death_handler_found,
       'Expected NewRelic death_handler to NOT be registered when sidekiq.ignore_retry_errors is false'
-  end
-
-  def test_basic_job_execution_still_works
-    with_config(:'sidekiq.ignore_retry_errors' => true) do
-      segment = run_job
-
-      assert_predicate segment, :finished?
-      assert_predicate segment, :record_metrics?
-      assert segment.duration.is_a?(Float)
-    end
   end
 end
